refactor code base with utils.py and constants.py
