+++ d20_governance/game-configs/whimsy.yaml
@@ -19,22 +19,22 @@ game:
   - quit
 
   stages:
-      - name: "design-flag"
+      - stage: "design-flag"
         message: "**Your first task is to collaboratively design a flag for your community. Each member should propose one element (e.g., a color, shape, or symbol) to include on the flag. Once all the elements have been proposed, vote for your favorite using the ✅ emoji. The top three elements with the highest number of votes will be incorporated into the flag.**"
         action: "vote"
         timeout_mins: 4
 
-      - name: "establish-tradition"
+      - stage: "establish-tradition"
         message: "**Next, create a unique tradition for your community. Each member should suggest a quirky and imaginative tradition that the community can practice together. Once all suggestions have been submitted, vote for your favorite using the ✅ emoji. The tradition with the highest number of votes will be adopted by your community.**"
         action: "vote"
         timeout_mins: 4
 
-      - name: "select-spokescreature"
+      - stage: "select-spokescreature"
         message: "**Finally, it's time to select a spokescreature to represent your community in interactions with other whimsical realms. Propose various magical or mythical creatures and vote for your favorite using the ✅ emoji. The creature with the highest number of votes will become your community's spokescreature.**"
         action: "vote"
         timeout_mins: 5
 
-      - name: "obscure-patterns"
+      - stage: "obscure-patterns"
         message: "**The Obscurity:**\n\n
         As you dive further into this whimsical realm, you find yourself in a mysterious location beyond the grasp of the light. Perhaps nobody else has discovered it yet; or perhaps it is where the community convenes for secret meetings. Either way, you sense that something unusual is happening here."
         action: "obscure"
